move BaseSection and PageBanner & remove TwoCol
